UNI-39 - Cria CRUD de Categorias

+++ src/main/java/br/com/unifood/unifood/controller/UserController.java
@@ -1,12 +1,15 @@
 package br.com.unifood.unifood.controller;
 
-import br.com.unifood.unifood.model.Users;
+import br.com.unifood.unifood.model.users.Users;
 import br.com.unifood.unifood.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api/users")
@@ -26,8 +29,18 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable Long id) {
-        userService.deleteUserById(id);
+    public ResponseEntity<String> deleteUser(@PathVariable Long id) {
+        Optional<Users> user = userService.getUserById(id);
+        if(user.isEmpty()) {
+            return new ResponseEntity<>("Usuário não encontrado!", HttpStatus.NOT_FOUND);
+        } else {
+            userService.deleteUserById(id);
+            String names = user.stream() // Para pegar nome do usuário
+                    .map(Users::getName)
+                    .collect(Collectors.joining());
+
+            return new ResponseEntity<>("Usuario: " + names + " deletado com sucesso!",HttpStatus.OK);
+        }
     }
 
 }
